+++ src/main/java/com/elmakers/mine/bukkit/magic/command/MagicCommandExecutor.java
@@ -5,7 +5,6 @@ import com.elmakers.mine.bukkit.api.batch.SpellBatch;
 import com.elmakers.mine.bukkit.api.magic.Automaton;
 import com.elmakers.mine.bukkit.api.magic.Mage;
 import com.elmakers.mine.bukkit.api.magic.MagicAPI;
-import com.elmakers.mine.bukkit.api.spell.Spell;
 import com.elmakers.mine.bukkit.api.spell.SpellTemplate;
 import com.elmakers.mine.bukkit.api.wand.LostWand;
 import com.elmakers.mine.bukkit.api.wand.Wand;
@@ -111,9 +110,7 @@ public class MagicCommandExecutor extends MagicMapExecutor {
 			}
 			return true;
 		}
-		if (subCommand.equalsIgnoreCase("give") || subCommand.equalsIgnoreCase("sell")
-            || subCommand.equalsIgnoreCase("configure") || subCommand.equalsIgnoreCase("describe")
-            || subCommand.equalsIgnoreCase("check") || subCommand.equalsIgnoreCase("debug"))
+		if (subCommand.equalsIgnoreCase("give") || subCommand.equalsIgnoreCase("sell"))
 		{
 			Player player = null;
 			int argStart = 1;
@@ -152,22 +149,6 @@ public class MagicCommandExecutor extends MagicMapExecutor {
             {
                 boolean showWorth = subCommand.equalsIgnoreCase("sell");
                 return onMagicGive(sender, player, subCommand, args2);
-            }
-            if (subCommand.equalsIgnoreCase("check"))
-            {
-                return onMagicCheck(sender, player, args2);
-            }
-            if (subCommand.equalsIgnoreCase("debug"))
-            {
-                return onMagicDebug(sender, player, args2);
-            }
-            if (subCommand.equalsIgnoreCase("describe"))
-            {
-                return onMagicDescribe(sender, player, args2);
-            }
-            if (subCommand.equalsIgnoreCase("configure"))
-            {
-                return onMagicConfigure(sender, player, args2);
             }
 		}
         if (subCommand.equalsIgnoreCase("worth"))
@@ -782,10 +763,6 @@ public class MagicCommandExecutor extends MagicMapExecutor {
             addIfPermissible(sender, options, "Magic.commands.magic.", "worth");
             addIfPermissible(sender, options, "Magic.commands.magic.", "sell");
 			addIfPermissible(sender, options, "Magic.commands.magic.", "list");
-            addIfPermissible(sender, options, "Magic.commands.magic.", "describe");
-            addIfPermissible(sender, options, "Magic.commands.magic.", "configure");
-            addIfPermissible(sender, options, "Magic.commands.magic.", "check");
-            addIfPermissible(sender, options, "Magic.commands.magic.", "debug");
 		} else if (args.length == 2) {
 			if (args[0].equalsIgnoreCase("list")) {
 				addIfPermissible(sender, options, "Magic.commands.magic.list", "maps");
@@ -794,8 +771,6 @@ public class MagicCommandExecutor extends MagicMapExecutor {
 				addIfPermissible(sender, options, "Magic.commands.magic.list", "schematics");
 				addIfPermissible(sender, options, "Magic.commands.magic.list", "entities");
 				addIfPermissible(sender, options, "Magic.commands.magic.list", "tasks");
-			} if (args[0].equalsIgnoreCase("configure") || args[0].equalsIgnoreCase("describe") || args[0].equalsIgnoreCase("check") || args[0].equalsIgnoreCase("debug")) {
-                options.addAll(api.getPlayerNames());
             } else if (args[0].equalsIgnoreCase("give") || args[0].equalsIgnoreCase("worth") || args[0].equalsIgnoreCase("sell")) {
 				options.add("wand");
 				options.add("material");
@@ -831,140 +806,4 @@ public class MagicCommandExecutor extends MagicMapExecutor {
 		}
 		return options;
 	}
-
-    public boolean onMagicCheck(CommandSender sender, Player player, String[] args)
-    {
-        Mage mage = api.getMage(player);
-		mage.debugPermissions(sender, null);
-        return true;
-    }
-
-    public boolean onMagicDebug(CommandSender sender, Player player, String[] args)
-    {
-        Mage mage = api.getMage(player);
-        if (args.length > 0) {
-            try {
-                int level = Integer.parseInt(args[0]);
-                mage.setDebugLevel(level);
-				if (level > 0) {
-					mage.setDebugger(sender);
-				} else {
-					mage.setDebugger(null);
-				}
-                sender.sendMessage(ChatColor.GOLD + "Setting debug level for  " + ChatColor.AQUA + player.getDisplayName() + ChatColor.GOLD + " to " + ChatColor.GREEN + Integer.toString(level));
-            } catch (Exception ex) {
-                sender.sendMessage("Expecting integer, got: " + args[0]);
-            }
-            return true;
-        }
-        if (mage.getDebugLevel() > 0) {
-            sender.sendMessage(ChatColor.GOLD + "Disabling debug for  " + ChatColor.AQUA + player.getDisplayName());
-            mage.setDebugLevel(0);
-			mage.setDebugger(null);
-        } else {
-            sender.sendMessage(ChatColor.AQUA + "Enabling debug for  " + ChatColor.AQUA + player.getDisplayName());
-            mage.setDebugLevel(1);
-			mage.setDebugger(sender);
-        }
-        return true;
-    }
-
-    public boolean onMagicDescribe(CommandSender sender, Player player, String[] args)
-    {
-        Mage mage = api.getMage(player);
-        ConfigurationSection data = mage.getData();
-        if (args != null && args.length > 0)
-        {
-            if (args[0].equals("*"))
-            {
-                sender.sendMessage(ChatColor.GOLD + "Mage data for " + ChatColor.AQUA + player.getDisplayName() + ChatColor.GOLD + ": ");
-                Collection<Spell> spells = mage.getSpells();
-                if (spells.size() == 0) {
-                    sender.sendMessage(ChatColor.RED + "No spell casts!");
-                    return true;
-                }
-                for (Spell spell : spells) {
-                    sender.sendMessage(ChatColor.LIGHT_PURPLE + spell.getName() + ChatColor.AQUA + " Cast Count: " + ChatColor.GOLD + spell.getCastCount());
-                }
-                return true;
-            }
-            Spell spell = mage.getSpell(args[0]);
-            if (spell != null)
-            {
-                sender.sendMessage(ChatColor.GOLD + "Mage data for " + ChatColor.AQUA + player.getDisplayName() + ChatColor.GOLD + ": " + ChatColor.LIGHT_PURPLE + spell.getName());
-                sender.sendMessage(ChatColor.AQUA + " Cast Count: " + ChatColor.GOLD + spell.getCastCount());
-                return true;
-            }
-            ConfigurationSection subSection = data.getConfigurationSection(args[0]);
-            if (subSection == null) {
-                sender.sendMessage(ChatColor.RED + "Unknown subsection: " + args[0]);
-                return true;
-            }
-            data = subSection;
-        }
-        Collection<String> keys = data.getKeys(false);
-        sender.sendMessage(ChatColor.GOLD + "Mage data for " + ChatColor.AQUA + player.getDisplayName());
-        for (String key : keys) {
-            if (data.isConfigurationSection(key)) {
-                ConfigurationSection subSection = data.getConfigurationSection(key);
-                sender.sendMessage(ChatColor.AQUA + " " + key + ChatColor.DARK_AQUA + " (" + subSection.getKeys(true).size() + " items)");
-            } else {
-                String value = data.getString(key);
-                if (value != null) {
-                    sender.sendMessage(ChatColor.AQUA + " " + key + ChatColor.DARK_AQUA + " (" + value + ")");
-                } else {
-                    sender.sendMessage(ChatColor.AQUA + " " + key);
-                }
-            }
-        }
-        return true;
-    }
-
-    public boolean onMagicConfigure(CommandSender sender, Player player, String[] args)
-    {
-        if (args.length != 2)
-        {
-            sender.sendMessage(ChatColor.RED + "Usage: " + ChatColor.AQUA + "magic configure <player> <key> <value>");
-            return true;
-        }
-        Mage mage = api.getMage(player);
-        if (args[0].equals("*"))
-        {
-            long value = 0;
-            try {
-                value = Long.parseLong(args[1]);
-            } catch (Exception ex) {
-                sender.sendMessage(ChatColor.RED + "Cast count must be a number");
-                return true;
-            }
-            Collection<Spell> spells = mage.getSpells();
-            for (Spell spell : spells)
-            {
-                spell.setCastCount(value);
-            }
-            sender.sendMessage(ChatColor.GOLD + "Set all spell cast counts to " + ChatColor.AQUA + value + ChatColor.GOLD + " for " + ChatColor.DARK_AQUA + player.getDisplayName());
-            return true;
-        }
-        Spell spell = mage.getSpell(args[0]);
-        if (spell != null)
-        {
-            long value = 0;
-            try {
-                value = Long.parseLong(args[1]);
-            } catch (Exception ex) {
-                sender.sendMessage(ChatColor.RED + "Cast count must be a number");
-                return true;
-            }
-            spell.setCastCount(value);
-            sender.sendMessage(ChatColor.GOLD + "Set " + ChatColor.AQUA + spell.getName() + ChatColor.GOLD + " cast count to " + ChatColor.AQUA + value + ChatColor.GOLD + " for " + ChatColor.DARK_AQUA + player.getDisplayName());
-            return true;
-        }
-
-        ConfigurationSection data = mage.getData();
-        String key = args[0];
-        String value = args[1];
-        data.set(key, value);
-        sender.sendMessage(ChatColor.GOLD + "Set " + ChatColor.AQUA + key + ChatColor.GOLD + " to " + ChatColor.AQUA + value + ChatColor.GOLD + " for " + ChatColor.DARK_AQUA + player.getDisplayName());
-        return true;
-    }
 }